Add files via upload
These screenshoots are important. When I was originally given the code, there were no alt tags to describe the images if they were not to show. In addition, there were far too many div tags when I first opened the code. I had to change many of the to "aside," "article," "footer","section" and many other tags that were more useful to making the website function. Lastly, it is important that you see the id element before Search engine optimization. That was not present when I first opened the code. This is important because without this the nav bar for Search Engine Optimization, would move to the Seo section of the page when clicked.